backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart

openai
python-dotenv
python-docx
reportlab
